RPC: Add locking RPC to all-clusters ESP app

+++ examples/common/pigweed/rpc_services/Device.h
@@ -84,7 +84,7 @@ public:
             response.product_id = static_cast<uint32_t>(product_id);
         }
 
-        uint32_t software_version;
+        uint16_t software_version;
         if (DeviceLayer::ConfigurationMgr().GetFirmwareRevision(software_version) == CHIP_NO_ERROR)
         {
             response.software_version = software_version;
